Move all timestamps to int

+++ Source/DSP/Grain.h
@@ -15,7 +15,7 @@
 class Grain {
  public:
   Grain() : duration(0), pbRate(1.0), startPos(0), trigTs(0), gain(0.0), pan(0.0) {}
-  Grain(std::vector<float> env, int duration_, float pbRate_, int startPos_, long trigTs_, float gain_, float pan_)
+  Grain(std::vector<float> env, int duration_, float pbRate_, int startPos_, int trigTs_, float gain_, float pan_)
       : duration(duration_),
         pbRate(pbRate_),
         startPos(juce::jmax(0, startPos_)),
@@ -24,12 +24,12 @@ class Grain {
         pan(pan_),
         mEnv(env) {}
 
-  float process(float chanPerc, const juce::AudioBuffer<float>& audioBuffer, float gain, long time);
+  float process(float chanPerc, const juce::AudioBuffer<float>& audioBuffer, float gain, int time);
 
   const int duration;  // Grain duration in samples
   const float pbRate;  // Playback rate (1.0 being regular speed)
   const int startPos;  // Start position in file to play from in samples
-  const long trigTs;   // Timestamp when grain was triggered in samples
+  const int trigTs;   // Timestamp when grain was triggered in samples
   const float gain;  // Grain gain
   const float pan;
